ocr-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
easyocr
torch
torchvision
streamlit
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0